connection-with-permission

+++ src/main/java/com/services/group4/snippet/services/SnippetService.java
@@ -1,13 +1,20 @@
 package com.services.group4.snippet.services;
 
 import com.services.group4.snippet.common.Language;
+import com.services.group4.snippet.dto.AllSnippetResponseDto;
+import com.services.group4.snippet.dto.ResponseDto;
 import com.services.group4.snippet.dto.SnippetDto;
 import com.services.group4.snippet.dto.SnippetResponseDto;
 import com.services.group4.snippet.model.Snippet;
 import com.services.group4.snippet.repositories.SnippetRepository;
+
+import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -27,43 +34,82 @@ public class SnippetService {
     this.permissionService = permissionService;
   }
 
-  public SnippetResponseDto createSnippet(SnippetDto snippetDto, Long userId) {
+  public ResponseEntity<ResponseDto<SnippetResponseDto>> createSnippet(SnippetDto snippetDto, String username, String userId) {
     Language language = new Language(snippetDto.getLanguage(), snippetDto.getVersion());
-    Snippet snippet = new Snippet(snippetDto.getName(), userId, language);
+    Snippet snippet = new Snippet(snippetDto.getName(), username, language);
 
     snippetRepository.save(snippet);
 
-    blobStorageService.saveSnippet(container, snippet.getId(), snippetDto.getContent());
+    // TODO: save snippet content from blob storage from infra bucket
+    //blobStorageService.saveSnippet(container, snippet.getId(), snippetDto.getContent());
 
-    permissionService.grantOwnerPermission(snippet.getId());
-    return new SnippetResponseDto(
-        snippet.getId(), snippet.getName(), snippetDto.getContent(), snippet.getLanguage());
+    ResponseEntity<ResponseDto<Long>> response = permissionService.grantOwnerPermission(snippet.getId(), userId);
+
+    if (response.getStatusCode().equals(HttpStatus.CREATED)) {
+      return new ResponseEntity<>(new ResponseDto<>( "Snippet created successfully",
+          new SnippetResponseDto(snippet.getId(), snippet.getName(), snippetDto.getContent(), snippet.getLanguage())),
+          HttpStatus.CREATED);
+    }
+    snippetRepository.delete(snippet);
+    return new ResponseEntity<>(new ResponseDto<>( "Something went wrong creating the snippet", null),
+        HttpStatus.INTERNAL_SERVER_ERROR);
   }
 
-  public Optional<SnippetResponseDto> getSnippet(Long snippetId) {
+  public ResponseEntity<ResponseDto<SnippetResponseDto>> getSnippet(Long snippetId, String userId) {
     Optional<Snippet> snippetOptional = this.snippetRepository.findSnippetById(snippetId);
 
     if (snippetOptional.isEmpty()) {
-      throw new NoSuchElementException("Snippet not found");
+      return new ResponseEntity<>(new ResponseDto<>("Snippet not found", null), HttpStatus.NOT_FOUND);
     }
+
+    ResponseEntity<ResponseDto<Boolean>> hasPermission = permissionService.hasPermissionOnSnippet(userId, snippetId);
+
+    if (Objects.requireNonNull(hasPermission.getBody()).data() != null && !hasPermission.getBody().data()) {
+      return new ResponseEntity<>(new ResponseDto<>("User does not have permission to view snippet", null), HttpStatus.FORBIDDEN);
+    }
+
     Snippet snippet = snippetOptional.get();
 
-    Optional<String> content = blobStorageService.getSnippet(container, snippetId);
+    // TODO: get snippet content from blob storage from infra bucket
+    Optional<String> content = "contenttttt".describeConstable(); //blobStorageService.getSnippet(container, snippetId);
 
     if (content.isEmpty()) {
-      throw new NoSuchElementException("Snippet content not found");
+      return new ResponseEntity<>(new ResponseDto<>("Snippet content not found", null), HttpStatus.NOT_FOUND);
     }
 
-    return Optional.of(
-        new SnippetResponseDto(
-            snippet.getId(), snippet.getName(), content.get(), snippet.getLanguage()));
+    return new ResponseEntity<>(new ResponseDto<>("Snippet found successfully",
+        new SnippetResponseDto(snippet.getId(), snippet.getName(), content.get(), snippet.getLanguage())),
+        HttpStatus.OK);
   }
 
-  public Optional<SnippetResponseDto> updateSnippet(Long id, SnippetDto snippetRequest) {
+  // este no va a tener el content del snippet solo la data de la tabla para la UI
+  public ResponseEntity<ResponseDto<List<AllSnippetResponseDto>>> getAllSnippet(String userId) {
+    ResponseEntity<ResponseDto<List<Long>>> snippetIds = permissionService.getAllowedSnippets(userId);
+
+    if (snippetIds.getStatusCode().isError() || snippetIds.getBody() == null) {
+      throw new SecurityException("User does not have permission to view snippets, because it has no snippets");
+    }
+
+    List<AllSnippetResponseDto> snippets =  snippetIds.getBody().data().stream()
+        .map(snippetId -> snippetRepository.findSnippetById(snippetId)
+            .map(snippet -> new AllSnippetResponseDto(snippet.getId(), snippet.getName(), snippet.getLanguage())))
+        .filter(Optional::isPresent)
+        .map(Optional::get)
+        .toList();
+
+    return new ResponseEntity<>(new ResponseDto<>("All snippets that has permission", snippets), HttpStatus.OK);
+}
+
+  public ResponseEntity<ResponseDto<SnippetResponseDto>> updateSnippet(Long id, SnippetDto snippetRequest, String userId) {
     Optional<Snippet> snippetOptional = snippetRepository.findById(id);
 
     if (snippetOptional.isEmpty()) {
       throw new NoSuchElementException("Snippet not found");
+    }
+
+    ResponseEntity<ResponseDto<Boolean>> hasPermission = permissionService.hasPermissionOnSnippet(userId, id);
+    if (Objects.requireNonNull(hasPermission.getBody()).data() != null && !hasPermission.getBody().data()) {
+      return new ResponseEntity<>(new ResponseDto<>(hasPermission.getBody().message(), null), HttpStatus.FORBIDDEN);
     }
 
     Snippet snippet = snippetOptional.get();
@@ -71,27 +117,38 @@ public class SnippetService {
     Language language = new Language(snippetRequest.getLanguage(), snippetRequest.getVersion());
     snippet.setLanguage(language);
 
-    blobStorageService.saveSnippet(container, snippet.getId(), snippetRequest.getContent());
+    // TODO: update snippet content from blob storage infra bucket
+    //blobStorageService.saveSnippet(container, snippet.getId(), snippetRequest.getContent());
 
     snippetRepository.save(snippet);
 
-    return Optional.of(
-        new SnippetResponseDto(
-            snippet.getId(),
-            snippet.getName(),
-            snippetRequest.getContent(),
-            snippet.getLanguage()));
+    return new ResponseEntity<>(new ResponseDto<>("Snippet updated successfully",
+        new SnippetResponseDto(snippet.getId(), snippet.getName(), snippetRequest.getContent(), snippet.getLanguage())),
+        HttpStatus.OK);
   }
 
-  public void deleteSnippet(Long id) {
+  public ResponseEntity<ResponseDto<Long>> deleteSnippet(Long id, String userId) {
     Optional<Snippet> snippetOptional = snippetRepository.findById(id);
 
     if (snippetOptional.isEmpty()) {
-      throw new NoSuchElementException("Snippet not found");
+      return new ResponseEntity<>(new ResponseDto<>("Snippet not found", id), HttpStatus.NOT_FOUND);
     }
 
     Snippet snippet = snippetOptional.get();
+
+    // TODO: delete snippet content from blob storage from infra bucket
+    //blobStorageService.deleteSnippet(container, snippet.getId());
+
+    ResponseEntity<ResponseDto<Long>> responseOwnership = permissionService.deletePermissions(id, userId);
+    if (responseOwnership.getStatusCode().isError()) {
+      return new ResponseEntity<>(new ResponseDto<>(Objects.requireNonNull(responseOwnership.getBody()).message(), id), responseOwnership.getStatusCode());
+    }
+
     snippetRepository.delete(snippet);
-    blobStorageService.deleteSnippet(container, snippet.getId());
+    return new ResponseEntity<>(new ResponseDto<>("Snippet deleted", id), HttpStatus.OK);
+  }
+
+  public ResponseEntity<ResponseDto<Long>> shareSnippet(Long snippetId, String ownerId, String targetUserId) {
+    return permissionService.shareSnippet(snippetId, ownerId, targetUserId);
   }
 }
